requirements.txt
django>=4.0,<5.0
celery
redis
instagrapi
groq
python-dotenv
requests
